+++ source/idea/idea-data-model/src/ideadatamodel_meta/__init__.py
@@ -10,4 +10,4 @@
 #  and limitations under the License.
 
 __name__ = 'idea-data-model'
-__version__ = '26.09.2'
+__version__ = '26.09.3'